Added the voice control-system for navigating the IPED-Toolkit. Source-Code is based on my Bachelor Thesis „Umsetzung und Evaluierung einer Sprachsteuerung zur Navigation in einer IVE“ 02/2015

+++ src/backend/js/views/LocationEditView.js
@@ -2,15 +2,18 @@
  * The iPED Toolkit
  * Backend
  *
- * (c) 2014 Tobias Brüggentisch, Morin Ostkamp
+ * (c) 2015 Tobias Brüggentisch, Morin Ostkamp, Nicholas Schiestel
  * Institute for Geoinformatics (ifgi), University of Münster
  */
 
 define(['backbonejs/js/backbone',
+        'backend/models/Locations',
+        'backend/models/Videos',
         'backend/models/Overlays',
-        'backend/models/Videos'
+        'backend/models/Relationship',
+        'bootstrap-bootbox/js/bootbox.min'
     ],
-    function(Backbone, Overlays, Videos) {
+    function(Backbone, Locations, Videos, Overlays, Relationship, bootbox) {
         /**
          * The backbone.js view used for editing a location
          */
@@ -19,6 +22,7 @@ define(['backbonejs/js/backbone',
                 this.backend = opts.backend;
                 this.title = opts.title;
                 this.isFetched = false;
+                this.relatedLocations = null;
                 this.videos = null;
                 this.overlays = null;
                 this.render();
@@ -42,6 +46,11 @@ define(['backbonejs/js/backbone',
                             thiz.$el.find('select[data-role=tagsinput]')
                                 .tagsinput('add', tag);
                         });
+                    thiz.$el.find('.bootstrap-tagsinput')
+                        .addClass('form-control')
+                        .css({
+                            height: "auto"
+                        });
                     thiz.fetch();
                 });
                 return this;
@@ -53,15 +62,21 @@ define(['backbonejs/js/backbone',
             fetch: function() {
                 var thiz = this;
 
-                if (this.isFetched == true) {
+                if (this.isFetched === true) {
                     return;
                 }
 
                 JL('iPED Toolkit.Backend')
                     .debug('Updating the LocationEditView');
+                JL('iPED Toolkit.Backend')
+                    .debug('Fetch videos in LocationEditView');
                 this.videos = new Videos();
                 this.videos.fetch({
                     success: function(model, response, options) {
+
+                        /*JL('iPED Toolkit.Backend')
+                            .debug(response);*/
+
                         thiz.$el.find('.videos')
                             .empty();
                         model.forEach(function(video) {
@@ -69,8 +84,10 @@ define(['backbonejs/js/backbone',
                             if (_.contains(thiz.model.get('videos'), video.get('id'))) {
                                 selected = 'selected';
                             }
+                            var option = '<option value="' + video.get('id') + '" ' + selected + '>' + video.get('name') + '</option>';
+
                             thiz.$el.find('.videos')
-                                .append('<option value="' + video.get('id') + '" ' + selected + '>' + video.get('name') + '</option>');
+                                .append(option);
                         });
                     },
                     error: function(model, response, options) {
@@ -79,9 +96,15 @@ define(['backbonejs/js/backbone',
                     }
                 });
 
+                JL('iPED Toolkit.Backend')
+                    .debug('Fetch overlays in LocationEditView');
                 this.overlays = new Overlays();
                 this.overlays.fetch({
                     success: function(model, response, options) {
+
+                        /*JL('iPED Toolkit.Backend')
+                            .debug(response);*/
+
                         thiz.$el.find('.overlays')
                             .empty();
                         model.forEach(function(overlay) {
@@ -99,11 +122,78 @@ define(['backbonejs/js/backbone',
                     }
                 });
 
+
+
+                JL('iPED Toolkit.Backend')
+                    .debug('Fetch relatedLocations in LocationEditView');
+
+                if (thiz.model.get('id') === undefined) {
+                    thiz.$el.find('.relatedLocations')
+                        .empty();
+                    thiz.$el.find('.relatedLocations')
+                        .html('<div id="textNone">None</div>');
+                } else {
+                    this.relatedLocations = new Locations();
+                    this.relatedLocations.url = '/api/locations/' + thiz.model.get('id') + '/locations';
+                    this.relatedLocations.fetch({
+                        success: function(model, response, options) {
+
+                            /*JL('iPED Toolkit.Backend')
+                                .debug(response);*/
+
+                            if (response.length === 0) {
+                                thiz.$el.find('.relatedLocations')
+                                    .empty();
+                                thiz.$el.find('.relatedLocations')
+                                    .html('<div id="textNone">None</div>');
+                            } else {
+
+                                thiz.$el.find('.relatedLocations')
+                                    .empty();
+                                thiz.$el.find('.relatedLocations')
+                                    .html('<table class="table" id="relatedLocationsTable' + thiz.model.get('id') + '"><tbody><tr><th></th><th>ID</th><th>Name</th><th>Relationship</th></tr></tbody></table>');
+                                model.forEach(function(relatedLocation) {
+
+                                    relatedLocation.relationship = new Relationship();
+                                    relatedLocation.relationship.url = '/api/locations/' + thiz.model.get('id') + '/locations/' + relatedLocation.get('id');
+
+                                    relatedLocation.relationship.fetch({
+                                        success: function(model, response, options) {
+
+                                            relatedLocation.relationship.url = '/api/relationships/' + relatedLocation.relationship.id;
+
+                                            thiz.$el.find('#relatedLocationsTable' + thiz.model.get('id') + ' > tbody:last')
+                                                .append('<tr><td><input type="radio" value="' + relatedLocation.relationship.id + '" name="relationship" class="_relationship"></td><td>' + relatedLocation.get('id') + '</td><td>' + relatedLocation.get('name') + '</td><td>' + relatedLocation.relationship.id + '</td></tr>');
+                                        },
+                                        error: function(model, response, options) {
+                                            JL('iPED Toolkit.Backend')
+                                                .error(response);
+                                        }
+                                    });
+
+                                });
+
+                            }
+                        },
+                        error: function(model, response, options) {
+                            JL('iPED Toolkit.Backend')
+                                .error(response);
+                        }
+                    });
+                }
+
+
                 this.isFetched = true;
             },
             events: {
                 'click button.close': '_close',
+                'click button.cancel': '_close',
                 'click button.save': '_save',
+                'click button.edit-relationship': '_editRelationship',
+                'click button.delete-relationship': '_deleteRelationship',
+                'click button.add-video': '_addVideo',
+                'click button.edit-video': '_editVideo',
+                'click button.delete-video': '_deleteVideo',
                 'click button.add-overlay': '_addOverlay',
                 'click button.edit-overlay': '_editOverlay',
                 'click button.delete-overlay': '_deleteOverlay'
@@ -122,14 +212,122 @@ define(['backbonejs/js/backbone',
             },
             _save: function() {
                 this._disableButtons();
+
                 this.backend.saveLocation({
                     location: this.model,
                     attributes: this.backend.form2js(this.$el.find('form')[0], '.', true),
                     dialog: this
                 });
             },
+            _editRelationship: function() {
+
+                var relationshipId = this.$el.find('._relationship:radio:checked')
+                    .attr('value');
+
+                if (relationshipId !== undefined) {
+
+                    var relationship = null;
+
+                    for (var i = 0; i < this.relatedLocations.models.length; i++) {
+
+                        if (this.relatedLocations.models[i].relationship.id == relationshipId) {
+                            relationship = this.relatedLocations.models[i].relationship;
+
+                            this.backend.editRelationship({
+                                relationship: relationship
+                            });
+                        }
+                    }
+                }
+            },
+            _deleteRelationship: function() {
+
+                var thiz = this;
+
+                var relationshipId = this.$el.find('._relationship:radio:checked')
+                    .attr('value');
+
+                if (relationshipId !== undefined) {
+
+                    var question = 'Are you sure you want to delete this relationship: <b>' + relationshipId + '</b>?';
+                    bootbox.dialog({
+                        title: "Attention",
+                        message: question,
+                        buttons: {
+                            cancel: {
+                                label: "Cancel",
+                                className: "btn-default",
+                                callback: function() {}
+                            },
+                            delete: {
+                                label: "OK",
+                                className: "btn-primary",
+                                callback: function() {
+
+                                }
+                            }
+                        }
+                    });
+                }
+            },
+            _addVideo: function() {
+                this.backend.addVideo();
+            },
+            _editVideo: function() {
+                if (this.$el.find('.videos option:selected')
+                    .attr('value') == -1) {
+                    this.$el.find('.videos')
+                        .parentsUntil('div.form-group')
+                        .addClass('has-error');
+                    alert('Please select a video footage first.');
+                    return;
+                }
+
+                var videoId = this.$el.find('.videos option:selected')
+                    .attr('value');
+
+                var video = this.videos.get(videoId);
+
+                console.warn(video);
+
+                this.backend.editVideo({
+                    video: video
+                });
+            },
+            _deleteVideo: function() {
+
+                var thiz = this;
+
+                var videoId = this.$el.find('.videos option:selected')
+                    .attr('value');
+                var video = this.videos.get(videoId);
+
+                // Confirmation before deleting
+                var question = 'Are you sure you want to delete the video: <b>' + video.attributes.name + '</b>?';
+                bootbox.dialog({
+                    title: "Attention",
+                    message: question,
+                    buttons: {
+                        cancel: {
+                            label: "Cancel",
+                            className: "btn-default",
+                            callback: function() {}
+                        },
+                        delete: {
+                            label: "OK",
+                            className: "btn-primary",
+                            callback: function() {
+                                thiz.backend.deleteVideo({
+                                    video: video
+                                });
+                            }
+                        }
+                    }
+                });
+
+            },
             _addOverlay: function() {
-                if (this.$el.find('.videos :selected')
+                if (this.$el.find('.videos option:selected')
                     .attr('value') == -1) {
                     this.$el.find('.videos')
                         .parentsUntil('div.form-group')
@@ -141,7 +339,7 @@ define(['backbonejs/js/backbone',
                     .parentsUntil('div.form-group')
                     .removeClass('has-error');
 
-                var videoId = this.$el.find('.videos :selected')
+                var videoId = this.$el.find('.videos option:selected')
                     .attr('value');
                 var video = this.videos.get(videoId);
                 this.backend.addOverlay({
@@ -149,7 +347,7 @@ define(['backbonejs/js/backbone',
                 });
             },
             _editOverlay: function() {
-                if (this.$el.find('.videos :selected')
+                if (this.$el.find('.videos option:selected')
                     .attr('value') == -1) {
                     this.$el.find('.videos')
                         .parentsUntil('div.form-group')
@@ -161,8 +359,8 @@ define(['backbonejs/js/backbone',
                     .parentsUntil('div.form-group')
                     .removeClass('has-error');
 
-                if (this.$el.find('.overlays :selected')
-                    .length == 0) {
+                if (this.$el.find('.overlays option:selected')
+                    .length === 0) {
                     this.$el.find('.overlays')
                         .parentsUntil('div.form-group')
                         .addClass('has-error');
@@ -173,10 +371,10 @@ define(['backbonejs/js/backbone',
                     .parentsUntil('div.form-group')
                     .removeClass('has-error');
 
-                var videoId = this.$el.find('.videos :selected')
+                var videoId = this.$el.find('.videos option:selected')
                     .attr('value');
                 var video = this.videos.get(videoId);
-                var overlayId = this.$el.find('.overlays :selected')
+                var overlayId = this.$el.find('.overlays option:selected')
                     .attr('value');
                 var overlay = this.overlays.get(overlayId);
                 this.backend.editOverlay({
@@ -185,8 +383,11 @@ define(['backbonejs/js/backbone',
                 });
             },
             _deleteOverlay: function() {
-                if (this.$el.find('.overlays :selected')
-                    .length == 0) {
+
+                var thiz = this;
+
+                if (this.$el.find('.overlays option:selected')
+                    .length === 0) {
                     this.$el.find('.overlays')
                         .parentsUntil('div.form-group')
                         .addClass('has-error');
@@ -197,11 +398,31 @@ define(['backbonejs/js/backbone',
                     .parentsUntil('div.form-group')
                     .removeClass('has-error');
 
-                var overlayId = this.$el.find('.overlays :selected')
+                var overlayId = this.$el.find('.overlays option:selected')
                     .attr('value');
                 var overlay = this.overlays.get(overlayId);
-                this.backend.deleteOverlay({
-                    overlay: overlay
+
+                // Confirmation before deleting
+                var question = 'Are you sure you want to delete the overlay: <b>' + overlay.attributes.name + '</b>?';
+                bootbox.dialog({
+                    title: "Attention",
+                    message: question,
+                    buttons: {
+                        cancel: {
+                            label: "Cancel",
+                            className: "btn-default",
+                            callback: function() {}
+                        },
+                        delete: {
+                            label: "OK",
+                            className: "btn-primary",
+                            callback: function() {
+                                this.backend.deleteOverlay({
+                                    overlay: overlay
+                                });
+                            }
+                        }
+                    }
                 });
             }
         });
